added login create user modal and set up default user login page (still need to login when button clicked/Also refined personality page
