+++ client/src/pages/Login/Login.css
@@ -1,0 +1,7 @@
+.display-block {
+  display: block;
+}
+
+.display-none {
+  display: none;
+}
