+++ client/src/pages/PersonalityLand/PersonalityLandStyle.css
@@ -2,10 +2,30 @@
   text-align: center;
 }
 h1,
-p {
-  margin: 10px;
+p,
+img {
+  margin: 15px;
 }
-.container {
-  background-color: transparent;
-  margin: 2vh, 1% auto, 1vh;
+.personality-hedge {
+  margin-left: auto;
+  margin-right: auto;
+}
+#personality-container {
+  margin: 20px, 20px auto, 20px;
+  max-width: 1200px;
+  padding-top: 30px;
+  padding-bottom: 30px;
+}
+#hog-pic {
+  display: flex;
+  align-items: center;
+}
+.card-body,
+.card-text {
+  padding: 0px;
+  margin: 0px;
+  text-align: center;
+}
+.card-image-top {
+  margin-bottom: 1px;
 }
